ingestion/requirements.txt
pandas
db-dtypes
functions-framework>=3.5.0
google-cloud-bigquery>=3.25.0
